Refactor models into package
