+++ models/__init__.py
@@ -1,0 +1,33 @@
+from .auth import User, Role, Permission, role_permissions
+from .employees import Department, Employee, Dependent
+from .documents import DocumentType, Document
+from .attendance import Attendance
+from .leave import LeaveType, LeaveRequest, LeaveBalance
+from .timesheets import PayPeriod, Timesheet, TimeEntry
+from .payroll import Payroll, PayrollEntry
+from .compensation import (
+    SalaryStructure,
+    ComponentType,
+    IncentiveType,
+    SalaryComponent,
+    EmployeeCompensation,
+    EmployeeIncentive,
+    CompensationReport,
+)
+from .benefits import Benefit, EmployeeBenefit
+from .budget import Budget, BudgetItem, BudgetCategory
+
+__all__ = [
+    'User', 'Role', 'Permission', 'role_permissions',
+    'Department', 'Employee', 'Dependent',
+    'DocumentType', 'Document',
+    'Attendance',
+    'LeaveType', 'LeaveRequest', 'LeaveBalance',
+    'PayPeriod', 'Timesheet', 'TimeEntry',
+    'Payroll', 'PayrollEntry',
+    'SalaryStructure', 'ComponentType', 'IncentiveType',
+    'SalaryComponent', 'EmployeeCompensation', 'EmployeeIncentive',
+    'CompensationReport',
+    'Benefit', 'EmployeeBenefit',
+    'Budget', 'BudgetItem', 'BudgetCategory',
+]
